1차 완성 기능 배포

Closes S09P12C210-175

See merge request s09-webmobile2-sub2/S09P12C210!139

+++ backend/src/main/java/com/kkini/core/domain/follow/service/FollowServiceImpl.java
@@ -5,9 +5,11 @@ import com.kkini.core.domain.follow.entity.Follow;
 import com.kkini.core.domain.follow.repository.FollowRepository;
 import com.kkini.core.domain.member.entity.Member;
 import com.kkini.core.domain.member.repository.MemberRepository;
+import com.kkini.core.domain.oauth2.UserPrincipal;
 import com.kkini.core.global.exception.NotFoundException;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
+import org.springframework.security.core.userdetails.User;
 import org.springframework.stereotype.Service;
 
 import javax.transaction.Transactional;
@@ -25,16 +27,23 @@ public class FollowServiceImpl implements FollowService{
     public void addFollow(FollowRequestDto followRequestDto) {
         Member me = memberRepository.findById(followRequestDto.getMemberId()).orElseThrow(() -> new NotFoundException(Member.class, followRequestDto.getMemberId()));
         Member target = memberRepository.findById(followRequestDto.getTargetMemberId()).orElseThrow(() -> new NotFoundException(Member.class, followRequestDto.getMemberId()));
-        followRepository.save(Follow.builder()
-                .me(me)
-                .target(target)
-                .build());
+        int count = followRepository.countByMe_IdAndTarget_Id(followRequestDto.getMemberId(), followRequestDto.getTargetMemberId());
+        if (count == 0){
+            followRepository.save(Follow.builder()
+                    .me(me)
+                    .target(target)
+                    .build());
+        }
     }
 
     @Override
-    public void deleteFollow(Long id) {
-        Follow follow = followRepository.findById(id).orElseThrow(() -> new NotFoundException(Follow.class));
-        followRepository.delete(follow);
+    public void deleteFollow(Long id, Long memberId) {
+        Follow follow = followRepository.findById(id).orElseThrow(() -> new NotFoundException(Follow.class ,id));
+        Member member = memberRepository.findById(memberId).orElseThrow(() -> new NotFoundException(Member.class, memberId));
+
+        if (follow.getMe().equals(member)){
+            followRepository.delete(follow);
+        }
     }
 
     @Override
